Fixed damage calculation issue
When the bullet's damage is zero (due to integer rounding / TTL timeout before object deletion) the damage of the bullet can be 0.
When calling decrease_health (in health_component.gd) there is an assertion that checks that the damage is greater than 0, to make sure signals are emitted only when health decreases and not stays constant.
Fixed this by adding a check before calling decrease_health that makes sure that the damage calculated after shields and such is greater than 0.